refactoring to submodule Copperegg::Alerts
