+++ lib/copperegg/alerts/schedule.rb
@@ -1,0 +1,64 @@
+require 'deep_merge'
+require 'copperegg/alerts/client'
+
+
+module Copperegg
+  module Alerts
+    class Schedule
+      attr_reader :schedules
+
+      def self.initialize
+        @client = Copperegg::Alerts::Client.instance
+        @schedules = @client.get('alerts/schedules.json')
+      end
+
+      def set(name, tags = {})
+        delete(name)
+        create(name, tags)
+      end
+
+      def modify(name, *args)
+        body = {}
+        args.each { |arg| body.deep_merge!(arg) }
+        selected_schedules = @schedules.select { |h| h['name'] == name }
+        if selected_schedules
+          @schedules -= selected_schedules
+          selected_schedules.each do |s|
+            result = @client.put?("alerts/schedules/#{s['id']}.json", body)
+            if result == nil
+              @schedules << s
+            else
+              @schedules << result
+            end
+          end
+        end
+      end
+
+      def create(name, *args)
+        defaults = {
+            'name' => name,
+            'state' => 'enabled',
+            'duration' => 10,
+            'start_time' => Time.now.gmtime.strftime('%Y-%m-%dt%H:%M:%Sz'),
+        }
+        args.each { |arg| defaults.deep_merge!(arg) }
+        if result = @client.post?('alerts/schedules.json', defaults)
+          @schedules << result.parsed_response
+        end
+      end
+
+      def delete(name)
+        selected_schedules = @schedules.select { |h| h['name'] == name }
+        if selected_schedules
+          @schedules -= selected_schedules
+          selected_schedules.each do |s|
+            if @client.delete?("alerts/schedules/#{s['id']}.json") == nil
+              @schedules << s
+            end
+          end
+        end
+      end
+    end
+  end
+end
+
