+++ lib/copperegg/alerts/version.rb
@@ -1,3 +1,5 @@
 module Copperegg
+  module Alerts
     VERSION = "0.0.2"
+  end
 end
